requirements.txt
coloredlogs==15.0.1
contourpy==1.1.0
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
flatbuffers==23.5.26
fonttools==4.40.0
humanfriendly==10.0
kiwisolver==1.4.4
matplotlib==3.7.1
mpmath==1.3.0
numpy==1.25.0
onnxruntime==1.15.1
packaging==23.1
Pillow==9.5.0
protobuf==4.23.3
pyparsing==3.1.0
python-dateutil==2.8.2
six==1.16.0
sympy==1.12
tk==0.1.0
